+++ demos/chinook/src/test/java/is/codion/framework/demos/chinook/model/InvoiceLineEditModelTest.java
@@ -74,7 +74,7 @@ public final class InvoiceLineEditModelTest {
 			theCallOfKtulu.put(Track.UNITPRICE, BigDecimal.valueOf(2));
 			theCallOfKtulu = connection.updateSelect(theCallOfKtulu);
 
-			editModel.set(invoiceLineBattery);
+			editModel.entity().set(invoiceLineBattery);
 			editModel.value(InvoiceLine.TRACK_FK).set(theCallOfKtulu);
 			editModel.update();