+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityDialogs.java
@@ -700,7 +700,7 @@ public final class EntityDialogs {
 
 		private void initializeEditModel(SwingEntityEditModel editModel) {
 			if (entity != null) {
-				editModel.set(entity.get());
+				editModel.entity().set(entity.get());
 			}
 			else {
 				editModel.revert();