+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - ColumnConditionPanel.focusGainedEvent() renamed focusGainedObserver().
 ### is.codion.framework.model
 - EntityEditModel.valueEdited() renamed edited().
+- EntityEditModel.entity() now Mutable based.
 ### is.codion.swing.framework.ui
 - EntityTablePanel bug fixed, Config.SHOW_REFRESH_PROGRESS_BAR now used for default config value.
 - EntityDialogs bug fixed, now respects the location if specified.